refactor: install integration field components refactor: required custom fields
