+++ src/components/Configure/types.ts
@@ -1,0 +1,20 @@
+import {
+  HydratedIntegrationField,
+  HydratedIntegrationFieldExistent,
+  IntegrationFieldMapping,
+} from '../../services/api';
+
+export type ConfigureStateIntegrationField = HydratedIntegrationFieldExistent & {
+  value: string | number | boolean | null,
+};
+
+export type CustomConfigureStateIntegrationField = IntegrationFieldMapping & {
+  value: string | number | undefined,
+};
+
+export type ConfigureState = {
+  allFields: HydratedIntegrationFieldExistent[] | null, // needed for custom mapping
+  requiredFields: HydratedIntegrationField[] | null,
+  optionalFields: ConfigureStateIntegrationField[] | null,
+  requiredCustomMapFields: CustomConfigureStateIntegrationField[] | null,
+};
