+++ src/components/Configure/utils.ts
@@ -75,7 +75,7 @@ export function getOptionalFieldsFromObject(object: HydratedIntegrationObject)
 export const getReadObject = (
   config: Config,
   objectName:string,
-) => config?.content?.read?.standardObjects[objectName];
+) => config?.content?.read?.standardObjects?.[objectName];
 
 // 5. get value for field
 export function getValueFromConfigExist(config: Config, objectName: string, key: string): boolean {
